Rename balance to mining rewards history
